Sat, Oct 14, 2023, 4:55 AM +05:00

+++ readability/readability.c
@@ -27,12 +27,12 @@ int main(void)
     printf("%i sentences\n", sentence_count);
 
     // Average number of letters per 100 words
-    int L = round((letter_count / (float) word_count) * 100);
-    printf("L: %i\n", L);
+    float L = round((letter_count / (float) word_count) * 100);
+    printf("L: %f\n", L);
 
     // Average number of sentences per 100 words
-    int S = round((sentence_count / (float) word_count) * 100);
-    printf("S: %i\n", S);
+    float S = round((sentence_count / (float) word_count) * 100);
+    printf("S: %f\n", S);
 
     // Coleman-Liau index formula
     int index = round(0.0588 * L - 0.296 * S - 15.8);
